// src/detran/models/wsib_models/debito.dto.ts
export class DebitoDTO {
    DescricaoServico: string;
    ValorAtualizadoFranquia: number;
    DataVencimento: string;
    DpvatCotas: string;
    IdDebito: number;
    Placa: string;
    IpvaExercicio: number;
    IpvaAnterior: number;
    LicenciamentoExercicio: number;
    LicenciamentoAnterior: number;
    TaxaServico: number;
    Multas: number;
    IpvaParcelamento: number;
    TaxaEspecial: number;
    TaxaPatio: number;
    DpvatExercicio: number;
    DpvatAnterior: number;
    CodigoServico: number;
    Classe: number;
    Exercicio: number;
    Parcela: number;
    IpvaCotas: string;
}